+++ Architectures/CleanArchitecture/Application/Sales/Queries/GetSaleDetail/SaleDetailModel.cs
@@ -16,10 +16,10 @@ namespace Application.Sales.Queries.GetSaleDetail
 
         public string ProductName { get; set; }
 
-        public decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
 
         public int Quantity { get; set; }
 
-        public decimal TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
     }
 }
